apc환경설정 품종, 규격 오류 조치, apc관리자 승인 개선

+++ src/main/java/com/at/apcss/am/cmns/vo/CmnsVrtyVO.java
@@ -33,7 +33,7 @@ public class CmnsVrtyVO extends CmnsItemVO {
 	/**
 	 * 중량단위
 	 */
-	private int unitWght;
+	private double unitWght;
 	/**
 	 * 중량계산구분코드
 	 */
